New rights licence and other updates...
In response to the remarks of Rintze:
Ok, I thought it might be better to use a more permissive license.
I've just realized that I have introduced a big bug with the management of pages in citations. I add the correction. Thank you a lot for the time you've spent with this style